Move launch focus logging into its own file
The IPC-probe fallback pushed launch.go past the 500-line architecture
limit, so the cohesive focus-logging block moves to launch_focus_log.go.
